# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/mfa/enroll": {
      "post": {
        "description": "Genera un QR code y secret para configurar en tu app de autenticación",
        "operationId": "MfaController_enrollMfa_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "QR code y secret generados",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "factorId": {
                      "type": "string"
                    },
                    "qrCode": {
                      "type": "string",
                      "description": "Base64 del QR code"
                    },
                    "secret": {
                      "type": "string",
                      "description": "Secret para configurar manualmente"
                    },
                    "uri": {
                      "type": "string",
                      "description": "URI para apps de autenticación"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Iniciar enrolamiento de 2FA",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/verify": {
      "post": {
        "description": "Verifica el código TOTP y activa 2FA para la cuenta",
        "operationId": "MfaController_verifyMfa_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA activado exitosamente"
          },
          "201": {
            "description": ""
          },
          "401": {
            "description": "Código inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verificar y activar 2FA",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/factors": {
      "get": {
        "description": "Obtiene la lista de métodos 2FA configurados para el usuario",
        "operationId": "MfaController_listFactors_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de factores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "friendlyName": {
                        "type": "string"
                      },
                      "factorType": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar factores MFA configurados",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/unenroll": {
      "delete": {
        "description": "Elimina un factor MFA de la cuenta",
        "operationId": "MfaController_unenrollMfa_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnenrollMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA deshabilitado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deshabilitar 2FA",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/status": {
      "get": {
        "description": "Indica si el usuario tiene 2FA habilitado y su nivel de autenticación",
        "operationId": "MfaController_getMfaStatus_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mfaEnabled": {
                      "type": "boolean"
                    },
                    "currentLevel": {
                      "type": "string",
                      "description": "aal1 = password, aal2 = password + MFA"
                    },
                    "factors": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verificar estado de 2FA",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/catalogs/unions": {
      "get": {
        "description": "Lista uniones de la organización, opcionalmente filtradas por país",
        "operationId": "CatalogsController_getUnions_v1",
        "parameters": [
          {
            "name": "countryId",
            "required": false,
            "in": "query",
            "description": "ID del país para filtrar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de uniones"
          }
        },
        "summary": "Obtener uniones",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/v1/catalogs/local-fields": {
      "get": {
        "description": "Lista campos locales, opcionalmente filtrados por unión",
        "operationId": "CatalogsController_getLocalFields_v1",
        "parameters": [
          {
            "name": "unionId",
            "required": false,
            "in": "query",
            "description": "ID de la unión para filtrar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de campos locales"
          }
        },
        "summary": "Obtener campos locales",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/v1/catalogs/districts": {
      "get": {
        "description": "Lista distritos, opcionalmente filtrados por campo local",
        "operationId": "CatalogsController_getDistricts_v1",
        "parameters": [
          {
            "name": "localFieldId",
            "required": false,
            "in": "query",
            "description": "ID del campo local para filtrar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de distritos"
          }
        },
        "summary": "Obtener distritos",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/v1/catalogs/churches": {
      "get": {
        "description": "Lista iglesias, opcionalmente filtradas por distrito",
        "operationId": "CatalogsController_getChurches_v1",
        "parameters": [
          {
            "name": "districtId",
            "required": false,
            "in": "query",
            "description": "ID del distrito para filtrar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de iglesias"
          }
        },
        "summary": "Obtener iglesias",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/v1/catalogs/roles": {
      "get": {
        "description": "Lista roles del sistema, opcionalmente filtrados por categoría (GLOBAL o CLUB)",
        "operationId": "CatalogsController_getRoles_v1",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Categoría de rol para filtrar",
            "schema": {
              "enum": [
                "GLOBAL",
                "CLUB"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de roles"
          }
        },
        "summary": "Obtener roles disponibles",
        "tags": [
          "catalogs"
        ]
      }
    },
...
    "/v1/catalogs/club-ideals": {
      "get": {
        "description": "Lista los ideales (ley, voto, lema, etc.) por tipo de club",
        "operationId": "CatalogsController_getClubIdeals_v1",
        "parameters": [
          {
            "name": "clubTypeId",
            "required": false,
            "in": "query",
            "description": "ID del tipo de club para filtrar",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ideales"
          }
        },
        "summary": "Obtener ideales de club",
        "tags": [
          "catalogs"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "factorId": {
            "type": "string",
            "example": "abc123-factor-id",
            "description": "ID del factor MFA a verificar"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "example": "123456",
            "description": "Código TOTP de 6 dígitos de tu app de autenticación"
          }
        },
        "required": [
          "factorId",
          "code"
        ]
      },
      "UnenrollMfaDto": {
        "type": "object",
        "properties": {
          "factorId": {
            "type": "string",
            "example": "abc123-factor-id",
            "description": "ID del factor MFA a eliminar"
          }
        },
        "required": [
          "factorId"
        ]
      },
...
    }
  }
}